refactor(#3275): decompose ensureNativeStringHelpers — extract String.prototype method helpers (slice 1)
Lift the cohesive String.prototype method-helper tail (search, trim,
pad/repeat, case, replace, split, construction, RegExp-escape) verbatim out
of the ~4.8k-LOC ensureNativeStringHelpers god-function in native-strings.ts
into three NEW cohesive sibling modules plus a shared-state factory:

- native-strings-shared.ts    — NativeStrShared bag + makeNativeStrShared
- native-strings-search.ts    — indexOf/lastIndexOf/includes/startsWith/endsWith + trim*
- native-strings-transform.ts — repeat/pad* + case + Unicode case + wellformed
- native-strings-rewrite.ts   — getSubstitution/replace*/split/fromChar*/regex_escape

The inline blocks are replaced by 8 builder calls in the original order,
AFTER the core helpers are registered (each looks them up by name in
ctx.nativeStrHelpers). Pure mechanical relocation — byte-identical emit
verified via prove-emit-identity (IDENTICAL, 39/39 gc/standalone/wasi).

native-strings.ts: 6,811 -> 4,062 LOC (-2,749). Every new module < 1500 LOC.

✓
